[move/#134] mapper 함수 파일 패키징
