+++ app/src/main/java/com/smashing/app/data/mapper/search/SuggestionItemMapper.kt
@@ -1,4 +1,4 @@
-package com.smashing.app.data.mapper
+package com.smashing.app.data.mapper.search
 
 import com.smashing.app.data.model.search.SuggestionItemModel
 import com.smashing.app.data.remote.dto.search.GetNicknameUsersSearchResponse
