+++ app/src/main/java/com/smashing/app/data/model/game/SubmissionConfirm.kt
@@ -1,0 +1,7 @@
+package com.smashing.app.data.model.game
+
+data class SubmissionConfirm(
+    val rating: String,
+    val content: String?,
+    val tags: List<String>?,
+)
