+++ app/src/main/java/com/smashing/app/data/remote/dto/game/PostConfirmSubmissionRequest.kt
@@ -1,0 +1,14 @@
+package com.smashing.app.data.remote.dto.game
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PostConfirmSubmissionRequest(
+    @SerialName("rating")
+    val rating: String,
+    @SerialName("content")
+    val content: String?,
+    @SerialName("tags")
+    val tags: List<String>?,
+)
